requirements.txt
streamlit
pandas
plotly
numpy
groq
scikit-learn
reportlab
pdfplumber
pillow
python-docx
openpyxl